feat(web): add YouTubeVideoRow component for video rows in channels and playlists
- Replaced custom video row implementation with `YouTubeVideoRow` component.
- Extracted video duration and view count formatting functions to `format.ts`.
- Updated page components to use `YouTubeVideoRow` for displaying videos.
