+++ web/src/routes/youtube/recent/+page.svelte
@@ -4,7 +4,7 @@
   import ArrowLeftRight from 'lucide-svelte/icons/arrow-left-right';
   import { getYouTubeRecentVideos, getYouTubeThumbnailUrl } from '$lib/api';
   import type { YoutubeVideo } from '$lib/api';
-  import { formatTimeAgo } from '$lib/youtube/format';
+  import { formatTimeAgo, formatDuration, formatViewCount } from '$lib/youtube/format';
   import { LoadedFade, YouTubeMediaRow } from '$lib/components/youtube';
 
   const DEFAULT_MAX_RESULTS = 25;
@@ -28,27 +28,6 @@
       sessionStorage.setItem('youtubeWatchReturnUrl', '/youtube/recent');
     }
     goto(`/youtube/watch/${encodeURIComponent(videoId)}`);
-  }
-
-  function formatViewCount(count: number): string {
-    if (count >= 1000000) {
-      return `${(count / 1000000).toFixed(1)}M views`;
-    }
-    if (count >= 1000) {
-      return `${(count / 1000).toFixed(1)}K views`;
-    }
-    return `${count} views`;
-  }
-
-  function formatDuration(seconds: number): string {
-    const hours = Math.floor(seconds / 3600);
-    const minutes = Math.floor((seconds % 3600) / 60);
-    const secs = seconds % 60;
-
-    if (hours > 0) {
-      return `${hours}:${minutes.toString().padStart(2, '0')}:${secs.toString().padStart(2, '0')}`;
-    }
-    return `${minutes}:${secs.toString().padStart(2, '0')}`;
   }
 </script>
